backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
tensorflow>=2.15.0
Pillow>=10.0.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0
